Remove dependency on Concat function.
The most common form of quicksort makes changes always on the
same underlying array, hence need of concatenation was an unnecessary
cost.

Removal of concat function meant that the sort function needed to
not return anything since all the changes are made directly on the
underlying array in all the operations. This change in sort() cascaded
in Sort() as well.