Fix a Major Bug in Quicksort - The pivot index was not moving
If there was any need in which pivot element needed to be move
left or right to rightly have the list partitioned at the pivot
index, then previously the pivot index was neither moving left
or right properly.

This commit also adds targetted tests to assert the movement of
the pivot index.

Please note we are talking only about the pivot index and not element
because once a list is partially compared against a pivot element,
you do want to complete the iteration and hence only change its index
and do partitions